refactor: Current Period Card
