+++ components/lookup/people/CurrentPeriodCard.tsx
@@ -1,0 +1,129 @@
+// Imports
+import cn from "@/utils/helpers/cn";
+import getLocaleString from "@/utils/helpers/getLocaleString";
+import getCurrentSchoolSessionState from "@/utils/helpers/schedule/getCurrentSchoolSessionState";
+import getTodaySetToPeriodTime from "@/utils/helpers/schedule/getTodaySetToPeriodTime";
+import useLocale from "@/utils/helpers/useLocale";
+import useNow from "@/utils/helpers/useNow";
+import { StylableFC } from "@/utils/types/common";
+import { SchedulePeriod } from "@/utils/types/schedule";
+import {
+  Card,
+  transition,
+  useAnimationConfig,
+} from "@suankularb-components/react";
+import { differenceInSeconds } from "date-fns";
+import { AnimatePresence, motion } from "framer-motion";
+import { useTranslation } from "next-i18next";
+import { useEffect, useState } from "react";
+
+/**
+ * A Card that contains the current period of a Student or Teacher and its
+ * progress in Search.
+ * 
+ * @param getCurrentPeriod A function that returns the current period of the Student or Teacher.
+ * @param onClick Triggers when the Card is clicked. Should open the Schedule.
+ * 
+ * @note
+ * Having a fetch function as a param like `getCurrentPeriod` is a bit
+ * experimental and has not been done in other components before. Let me know
+ * if you have any feedback.
+ */
+const CurrentPeriodCard: StylableFC<{
+  getCurrentPeriod: () => Promise<SchedulePeriod | null>;
+  onClick: () => void;
+}> = ({ getCurrentPeriod, onClick, style, className }) => {
+  const locale = useLocale();
+  const { t } = useTranslation("lookup", {
+    keyPrefix: "teachers.detail.teaching",
+  });
+
+  const now = useNow();
+
+  const { duration, easing } = useAnimationConfig();
+
+  const [loading, setLoading] = useState(true);
+  const [currentPeriod, setCurrentPeriod] = useState<SchedulePeriod | null>(
+    null,
+  );
+
+  /**
+   * The percentage of the current period that has passed.
+   */
+  const percentage = currentPeriod?.content.length
+    ? (differenceInSeconds(
+        now,
+        getTodaySetToPeriodTime(currentPeriod.start_time),
+      ) /
+        (currentPeriod.duration * 3000)) *
+      100
+    : 0;
+
+  // Fetch the current period once the percentage reaches 100%
+  useEffect(() => {
+    // If school is not in session, don’t fetch
+    if (getCurrentSchoolSessionState() !== "in-session") {
+      setLoading(false);
+      return;
+    }
+
+    // If the current period is already fetched and the percentage is 100%,
+    // don’t fetch
+    if (currentPeriod && percentage === 100) return;
+
+    // Fetch the current period
+    (async () => {
+      setLoading(true);
+      setCurrentPeriod(await getCurrentPeriod());
+      setLoading(false);
+    })();
+  }, [percentage === 100]);
+
+  return (
+    <Card
+      appearance="outlined"
+      stateLayerEffect
+      onClick={onClick}
+      style={style}
+      className={cn(
+        `relative isolate !m-[1px] min-h-[2.5rem] overflow-hidden rounded-md
+        !border-0 bg-surface px-3 py-2 hover:!m-0 hover:!border-1 focus:!m-0
+        focus:!border-1 focus:!border-outline-variant `,
+        loading && `animate-pulse`,
+        className,
+      )}
+    >
+      {/* Progress */}
+      <motion.div
+        initial={{ width: 0 }}
+        animate={{ width: !loading ? `${percentage}%` : "0%" }}
+        transition={transition(duration.medium2, easing.standard)}
+        className="absolute inset-0 right-auto -z-10 bg-surface-variant"
+      />
+
+      {/* Subject name / status */}
+      <AnimatePresence>
+        {!loading && (
+          <motion.p
+            initial={{ opacity: 0 }}
+            animate={{ opacity: 1 }}
+            exit={{ opacity: 0 }}
+            transition={transition(duration.short2, easing.standard)}
+            className="skc-text skc-text--title-medium"
+          >
+            {currentPeriod?.content.length
+              ? t("ongoing", {
+                  subject: getLocaleString(
+                    currentPeriod?.content[0].subject.name,
+                    locale,
+                  ),
+                })
+              : t("free")}
+          </motion.p>
+        )}
+      </AnimatePresence>
+    </Card>
+  );
+};
+
+export default CurrentPeriodCard;
